project/requirements_dev.txt
tensorflow-quantum==0.3.1
tensorflow-gpu==2.2.0
pytest==6.0.2
pytest-cov==2.10.1
pylint==2.6.0
pandas==1.1.2
matplotlib==3.3.2
jupyter==1.0.0
opencv-python==4.4.0.42
seaborn==0.11.0